# doc/api.yaml
openapi: 3.0.3
info:
  version: 'latest'
  title: FoodGen - API
  description: |
    A simple API that creates random food
security:
  - BearerAuth: []
paths:
...
  '/recipes/{id}':
    get:
      tags:
        - recipe
      summary: Get recipe by id
      operationId: getRecipeById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: return the specific recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
  '/recipes':
    get:
      tags:
        - recipe
      summary: Get all recipes
      operationId: getRecipes
      parameters:
        - name: page
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Page'
        - name: page_size
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/PageSize'
      responses:
        '200':
          description: a list of all recipes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - recipe
      summary: Create or update recipes
      operationId: crupdateRecipes
      requestBody:
        description: List of recipes to be crupdated
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/CreateRecipe'
      responses:
        '200':
          description: return a list crupdated recipes.
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Recipe'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestException'
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotAuthorizedException'
    '404':
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundException'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'
  schemas:
...
    CreateRecipe:
      type: object
      properties:
        name:
          type: string
          example: Pizza
        readme:
          type: string
          example: Bunch of text
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/RecipeIngredient'
    Recipe:
      allOf:
        - $ref: '#/components/schemas/CreateRecipe'
      type: object
      properties:
        id:
          type: string
          format: uuid
...
    CreateIngredient:
      type: object
      properties:
        name:
          type: string
          example: Tomato
...
    RecipeIngredient:
      allOf:
        - $ref: '#/components/schemas/CreateIngredient'
      type: object
      properties:
        measure:
          type: string
...
    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: InternalServerException
        message: Unexpected error